implemented past, future and today event distinctions and not being able to book a past event

+++ api/src/main/java/com/urban_events/api/service/BookingService.java
@@ -42,6 +42,10 @@ public class BookingService {
 
         Optional<Booking> existing = bookingRepository.findByUserAndEvent(user, event);
 
+        if (event.getEventDate().isBefore(LocalDateTime.now())) {
+            throw new RuntimeException("Cannot book a past event.");
+        }
+
         if (existing.isPresent()) {
             bookingRepository.delete(existing.get());
             return "unbooked";
